Extract external code element cache
